requirements.txt
appnope 
argon2-cffi 
async-generator==1.10
attrs 
backcall==0.2.0
bleach 
certifi==2021.10.8
cffi 
colorama 
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
importlib-metadata 
iniconfig==1.1.1
ipykernel 
ipython 
ipython-genutils==0.2.0
ipywidgets 
jedi 
Jinja2==2.11.2
jsonschema 
jupyter==1.0.0
jupyter-client 
jupyter-console 
jupyter-core==4.6.3
jupyterlab-pygments 
MarkupSafe 
mistune 
mkl-fft==1.2.0
mkl-random==1.2.0
mkl-service==2.3.0
mpi4py 
nbclient 
nbconvert 
nbformat 
nest-asyncio 
nfft==0.1
notebook 
numpy 
packaging==20.4
paddle-quantum
paddlepaddle-gpu
pandas 
pandocfilters==1.4.2
parso 
pexpect 
pickleshare 
pluggy==1.0.0
prometheus-client==0.8.0
prompt-toolkit 
ptyprocess==0.6.0
py==1.11.0
pycparser 
Pygments 
pyparsing==2.4.7
pyrsistent 
pytest==7.1.1
python-dateutil==2.8.1
pytz==2020.1
pyzmq==19.0.2
qtconsole 
QtPy==1.9.0
schwimmbad 
scipy 
Send2Trash==1.5.0
six==1.15.0
terminado==0.9.1
testpath==0.4.4
tomli==2.0.1
tornado==6.0.4
tqdm 
traitlets 
wcwidth 
webencodings==0.5.1
widgetsnbextension==3.5.1
zipp 
